fix(eve): compare dynamic skill marker keys safely
Guard same-size marker comparisons against disjoint package names and cover dynamic skill renames without leaving the marker or lifecycle half-committed. Keep recovery cases in a focused test module within Eve file-size law.
