+++ packages/eve/src/context/dynamic-skill-lifecycle.test.ts
@@ -237,86 +237,6 @@ describe("dispatchDynamicSkillEvent", () => {
     ).toBe(true);
   });
 
-  it.each(["missing", "stale"] as const)(
-    "invalidates unresolvable manifest packages after a %s marker",
-    async (markerState) => {
-      const { ctx, sandbox } = createCtx();
-      const sessionResolver = createResolver(
-        "session-policy",
-        () => makeSkill("Session policy", "Follow session policy."),
-        undefined,
-        ["session.started"],
-      );
-      const turnResolver = createResolver(
-        "turn-policy",
-        () => makeSkill("Turn policy", "Follow turn policy."),
-        undefined,
-        ["turn.started"],
-      );
-
-      await dispatchDynamicSkillEvent({
-        ctx,
-        event: makeEvent("session.started"),
-        messages: [],
-        resolvers: [sessionResolver, turnResolver],
-      });
-      await dispatchDynamicSkillEvent({
-        ctx,
-        event: makeEvent("turn.started"),
-        messages: [],
-        resolvers: [sessionResolver, turnResolver],
-      });
-      expect(ctx.get(PendingSkillAnnouncementKey)).toContain("session-policy: Session policy");
-
-      let activeSandbox = sandbox;
-      if (markerState === "missing") {
-        activeSandbox = mockSandbox({
-          id: "sbx_recreated",
-          commands: {
-            [HOME_PROBE_COMMAND]: { exitCode: 0, stderr: "", stdout: "/home/agent\n" },
-          },
-        });
-        ctx.set(SandboxKey, activeSandbox.access);
-      } else {
-        const markerPath = `/home/agent/.agents/skills/${DYNAMIC_SKILL_MATERIALIZATION_MARKER_FILE}`;
-        const marker = JSON.parse(sandbox.files.get(markerPath)!) as {
-          packages: Record<string, unknown>;
-          version: number;
-        };
-        delete marker.packages["session-policy"];
-        const staleMarker = JSON.stringify(marker);
-        sandbox.files.set(markerPath, staleMarker);
-        sandbox.fileBytes.set(markerPath, Buffer.from(staleMarker));
-      }
-
-      await dispatchDynamicSkillEvent({
-        ctx,
-        event: makeEvent("turn.started"),
-        messages: [],
-        resolvers: [sessionResolver, turnResolver],
-      });
-
-      expect(ctx.get(DynamicSkillManifestKey)).toEqual({
-        "turn-policy": [
-          {
-            contentDigest: expect.stringMatching(/^[a-f0-9]{64}$/u),
-            description: "Turn policy",
-            name: "turn-policy",
-            relativePaths: ["SKILL.md"],
-          },
-        ],
-      });
-      expect(ctx.get(PendingSkillAnnouncementKey)).not.toContain("session-policy: Session policy");
-      expect(ctx.get(PendingSkillAnnouncementKey)).toContain("turn-policy: Turn policy");
-      expect(activeSandbox.files.has("/home/agent/.agents/skills/session-policy/SKILL.md")).toBe(
-        false,
-      );
-      expect(activeSandbox.files.get("/home/agent/.agents/skills/turn-policy/SKILL.md")).toBe(
-        "Follow turn policy.",
-      );
-    },
-  );
-
   it.each([
     ["corrupt", "{"],
     ["old", JSON.stringify({ packages: {}, version: 0 })],
@@ -402,59 +322,6 @@ describe("dispatchDynamicSkillEvent", () => {
       "Policy body",
     );
     expect(sandbox.files.has(markerPath)).toBe(true);
-  });
-
-  it("clears stale siblings when retrying a partial first materialization", async () => {
-    const { ctx, sandbox } = createCtx();
-    let files: Readonly<Record<string, string>> = {
-      "references/stale.md": "Stale body",
-      "scripts/fail.sh": "exit 1",
-    };
-    const resolver = createResolver("tenant", () =>
-      makeSkill("Tenant policy", "Follow tenant policy.", files),
-    );
-    const failingSession = {
-      ...sandbox.session,
-      async writeBinaryFile(options: Parameters<typeof sandbox.session.writeBinaryFile>[0]) {
-        if (options.path.endsWith("/scripts/fail.sh")) {
-          throw new Error("injected final write failure");
-        }
-        await sandbox.session.writeBinaryFile(options);
-      },
-    };
-    ctx.set(SandboxKey, {
-      async captureState() {
-        return { initialized: false, session: null };
-      },
-      async get() {
-        return failingSession;
-      },
-    });
-
-    await expect(
-      dispatchDynamicSkillEvent({
-        ctx,
-        event: makeEvent(),
-        messages: [],
-        resolvers: [resolver],
-      }),
-    ).rejects.toThrow("injected final write failure");
-    expect(sandbox.files.has("/home/agent/.agents/skills/tenant/references/stale.md")).toBe(true);
-    expect(ctx.get(DynamicSkillManifestKey)).toBeUndefined();
-
-    files = {};
-    ctx.set(SandboxKey, sandbox.access);
-    await dispatchDynamicSkillEvent({
-      ctx,
-      event: makeEvent(),
-      messages: [],
-      resolvers: [resolver],
-    });
-
-    expect(sandbox.files.has("/home/agent/.agents/skills/tenant/references/stale.md")).toBe(false);
-    expect(sandbox.files.get("/home/agent/.agents/skills/tenant/SKILL.md")).toBe(
-      "Follow tenant policy.",
-    );
   });
 
   it("clears removed dynamic skills from the durable announcement", async () => {
